`beginPlay` -> `onLoadLibrary` (and friends)
Refactor and cleanup, two-parter. Removes the `beginPlay` function, since
that concept is better served in gamecode. In its stead, we now provide
an `onLoadLibrary` hook to allow the gamecode to (re)initialize any
buffers it requires when it's loaded. This function's contract requires
idempotency in implementations, but otherwise any logic the game library
needs is fair game.

Second part was a little function name clarification -- prefixing
library hooks with the word 'on' to denote that they are responding to
an event, not general-purpose functions. This means we now have:

 * `onStep` instead of `step`
 * `onLoadLibrary` instead of `beginPlay`

Finally, this also removes a bunch of code that's been commented out for
months in the implementation of `onLoadLibrary` (nee `beginPlay`).