Fix Ring iterators
The realization that stl iterators count end() as one element past the
given container (which will be a valid but non-dereferenceable pointer,
most times) spurred this change. Additional state is required for the
Ring::iterator now -- tracking how far the iterator has traversed
through the Ring. This may come in handy if we want bi-directional
iterators though, so it's not so bad.