Begin modifying the mem and state.memory APIs
This change set only modifies code that is not 1) test code, and 2)
buffer-relate code. The goal here was to settle on a sane, clear,
and discoverable nomenclature for the memory-related APIs. My
general decisions are as follows;

The mem:: namespace (and the associated state.memory fnptr API) will
remain distinct from the buffer:: namespace. The former is primarily
concerned with the mechanics of claiming, re-discovering, and
releasing memory. The later is designed to be a human-readable
wrapper around buffer operations.

The mem:: namespace exposes functions of the form `allocate`, `free`,
and `lookup`, all accepting and/or returning pointers (I suggest we
reason about these pointers as though they were `void*`s, rather than
`buffer*`s). The buffer:: namespace exposes functions of the form
`create`, `create_view`, `find_view`, etc., typically returning
`Optional`s that wrap buffers or buffer views.