Normalize collection initializers on bd vs m_bd
bd is the safer choice under refactoring since a change in member
declaration order can't throw the whole thing off. This changeset
applies that to all our other structures as well.

This change came from review comments on #287, which does it for
Ring along with some other member state changes.