Move from ANSI signal handlers to POSIX
This shift lets us use the signal context from the sa_sigaction variant
of the signal handler prototype, instead of just the signal number.

This'll let me rebase the trace on the instruction pointer when the signal was
caught, rather than starting from the far end of the signal trampoline.