Decided against a nonstd::memory:: namespace; it'd be ambiguous
I started moving array into nonstd::memory:: and realized there was
way too much crossover with with the ::memory:: cross-library. It
was getting confusing, so I decided to not do that.

Instead, move Buffer into the top-level nonstd:: namespace, and put
the type_id enum and the ResizeFn type into the Buffer struct as a
type members, for a poor-man's multipurpose namespace.