Improve cx_math.test.cc
This is largely a switch from `-std::numeric_limits<T>::max()` to
the more descriptive `std::numeric_limits<T>::lowest()`. And a few
renames and realignments to incorporate that change. (A "few"...)

The other big piece is correcting my construction of
`std::uniform_real_distribution<T>` objects. Previously I had been
using an out-of-specification range, which had been masking (what
is assumed to be) a rounding error in nonstd::cx::fmod.